Improve plots
- Show summed TBS of both transport blocks
- Troughput: SI conforming unit prefix (Mbit = 10^6 bit, instead of 2^20)
- Improve data aggregation for cell activity plot
  - sfn now uint32_t
  - tti integer overflow handling
  - Change plot title
- Code cleaning

+++ lib/include/falcon/CCSnifferInterfaces.h
@@ -101,7 +101,7 @@ public:
   uint32_t mcs_idx;
   int      mcs_tbs;
   uint32_t l_prb;
-  uint16_t sfn;
+  uint32_t sfn;
   uint32_t total_prb;
 };
